% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approve_assignment.R
\name{approve_assignment}
\alias{approve_assignment}
\title{Approve Assignment}
\usage{
approve_assignment(id, message, ..., .override = F)
}
\value{
If successful, Empty body and no errors.
}
\description{
Approves an assignment
}
