% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate_qualification_with_worker.R
\name{associate_qualification_with_worker}
\alias{associate_qualification_with_worker}
\title{Associate Qualification with Worker}
\usage{
associate_qualification_with_worker(type, id, value, ..., .notify = F)
}
\arguments{
\item{type}{The id of the Qualification type to use for the assigned Qualification.}

\item{id}{The id of the Worker to whom the Qualification is being assigned.}

\item{value}{The value of the Qualification to assign}

\item{...}{Placeholder for future arguments}

\item{.notify}{Specifies whether to send a notification email message to the Worker saying that the qualification was assigned to the Worker.}
}
\value{
If successful, empty body and no errors.
}
\description{
Associates a qualification with a worker
}
