% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_additional_assignments_for_hit.R
\name{create_additional_assignments_for_hit}
\alias{create_additional_assignments_for_hit}
\title{Create Additional Assignments for HIT}
\usage{
create_additional_assignments_for_hit(
  id,
  amount,
  ...,
  .token = rlang::missing_arg()
)
}
\value{
If successful, Empty body and no errors.
}
\description{
Creates additional assignments for a HIT
}
