% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hit.R
\name{create_hit}
\alias{create_hit}
\title{Create HIT}
\usage{
create_hit(
  title,
  description,
  questions,
  layout_id,
  layout_parameters,
  reward,
  assignment_duration,
  lifetime,
  keywords,
  max_assignments,
  auto_approval_delay,
  qualifications,
  assignment_review_policy,
  hit_review_policy,
  annotation,
  ...,
  .token
)
}
\arguments{
\item{.token}{}
}
\value{
If successful, a HIT object.
}
\description{
Creates a HIT
}
