% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hit_type.R
\name{create_hit_type}
\alias{create_hit_type}
\title{Create HIT Type}
\usage{
create_hit_type(
  title,
  description,
  reward,
  assignment_duration,
  keywords,
  auto_approval_delay,
  qualifications
)
}
\value{
If successful, a HIT type id.
}
\description{
Creates a HIT type
}
