% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hit_with_hit_type.R
\name{create_hit_with_hit_type}
\alias{create_hit_with_hit_type}
\title{Create HIT With HIT Type}
\usage{
create_hit_with_hit_type(
  type,
  question,
  layout_id,
  layout_parameters,
  lifetime,
  max_assignments,
  assignment_review_policy,
  hit_review_policy,
  annotation,
  ...,
  .token
)
}
\value{
If successful, a HIT object.
}
\description{
Creates a HIT of a particular HIT Type
}
