% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_qualification_type.R
\name{create_qualification_type}
\alias{create_qualification_type}
\title{Create Qualification Type}
\usage{
create_qualification_type(
  name,
  description,
  keywords,
  retry_delay,
  qualification_type_status,
  test,
  answer_key,
  test_duration,
  auto_granted,
  auto_granted_value
)
}
\arguments{
\item{name}{The name to be given to the Qualification type}

\item{description}{A long description for the Qualification type}

\item{keywords}{One or more words or phrases that describe the Qualification type}

\item{retry_delay}{The number of seconds that a Worker must wait after requesting a Qualification of the Qualification type before the worker can retry the Qualification request}

\item{qualification_type_status}{The initial status of the Qualification type}

\item{test}{The questions for the Qualification test a Worker must answer correctly to obtain a Qualification of this type; if this parameter is specified, \code{test_duration} must also be specified}

\item{answer_key}{The answers to the Qualification test specified in the Test parameter}

\item{test_duration}{The number of seconds the Worker has to complete the Qualification test, starting from the time the Worker requests the Qualification}

\item{auto_granted}{Specifies whether requests for the Qualification type are granted immediately, without prompting the Worker with a Qualification test}

\item{auto_granted_value}{The Qualification value to use for automatically granted Qualifications. This parameter is used only if the \code{auto_granted} parameter is true.}
}
\value{
If successful, a Qualification Type object.
}
\description{
Creates a Qualification Type
}
