% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_worker_block.R
\name{create_worker_block}
\alias{create_worker_block}
\title{Create Worker Block}
\usage{
create_worker_block(id, reason)
}
\arguments{
\item{id}{The id of the worker to block}

\item{reason}{A message that explains the reason for blocking the Worker. The Worker does not see this message}
}
\value{
If successful, an empty body with no errors
}
\description{
Blocks a Worker from working on your HITs
}
