% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_qualification_type.R
\name{delete_qualification_type}
\alias{delete_qualification_type}
\title{Delete Qualification Type}
\usage{
delete_qualification_type(id)
}
\arguments{
\item{id}{The id of the Qualification type}
}
\value{
If successful, empty body and no errors
}
\description{
Deletes a Qualification Type
}
