% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_worker_block.R
\name{delete_worker_block}
\alias{delete_worker_block}
\title{Delete Worker Block}
\usage{
delete_worker_block(id, reason)
}
\arguments{
\item{id}{The id of the Worker to unblock}

\item{reason}{A message that explains the reason for unblocking the Worker. The Worker does not see this message.}
}
\value{
If successful, empty body and no errors
}
\description{
Unblocks a Worker
}
