% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissociate_qualification_from_worker.R
\name{dissociate_qualification_from_worker}
\alias{dissociate_qualification_from_worker}
\title{Dissociate Qualification from Worker}
\usage{
dissociate_qualification_from_worker(id, worker, reason)
}
\arguments{
\item{id}{The id of the Qualification type of the Qualification to be revoked.}

\item{worker}{The id of the Worker who possesses the Qualification to be revoked.}

\item{reason}{A text message that explains why the Qualification was revoked. The user who had the Qualification sees this message. IF a reason is not provided, the worker will not be notified.}
}
\value{
If successful, empty body and no errors
}
\description{
Dissociates a Qualification from a Worker
}
