% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account_balance.R
\name{get_account_balance}
\alias{get_account_balance}
\title{Get Account Balance}
\usage{
get_account_balance()
}
\value{
If successful, a string representing your available balance details in U.S. Dollars
}
\description{
Gets the Account Balance
}
