% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assignment.R
\name{get_assignment}
\alias{get_assignment}
\title{Get Assignment}
\usage{
get_assignment(id)
}
\arguments{
\item{id}{The id of the assignment you want to retrieve}
}
\value{
If successful, an Assignment object
}
\description{
Gets the Assignment
}
