% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qualification_score.R
\name{get_qualification_score}
\alias{get_qualification_score}
\title{Get Qualification Score}
\usage{
get_qualification_score(qualification_type, worker)
}
\arguments{
\item{qualification_type}{The id of the Qualification Type}

\item{worker}{The id of the Worker whose Qualification is being updated}
}
\value{
If successful, a Qualification object
}
\description{
Gets the Value of a Worker's Qualification for a given Qualification type
}
