% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qualification_type.R
\name{get_qualification_type}
\alias{get_qualification_type}
\title{Get Qualification Type}
\usage{
get_qualification_type(id)
}
\arguments{
\item{id}{The id of the Qualification Type.}
}
\value{
If successful, a Qualification Type object
}
\description{
Gets information about a Qualification type using its id
}
