% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_assignments_for_hit.R,
%   R/list_bonus_payments.R
\name{list_assignments_for_hit}
\alias{list_assignments_for_hit}
\title{List Assignments for HIT}
\usage{
list_assignments_for_hit(id, assignment, next_token, max_results)

list_assignments_for_hit(id, assignment, next_token, max_results)
}
\arguments{
\item{id}{The id of the HIT}

\item{assignment}{The id of the assignment associated with the bonus payments to retrieve. If specified, only bonus payments for the given assignment are returned. Either the HITId parameter or the AssignmentId parameter must be specified}

\item{next_token}{Pagination token}

\item{max_results}{Maximum number of Assignments to return}

\item{assignment_statuses}{The status of the assignments to return: Submitted | Approved | Rejected}
}
\value{
If successful, a list of Assignment objects associated with a HIT

If successful, a list of Bonuses with the following fields: WorkerId, BonusAmount, AssignmentId, Reason and GrantTime
}
\description{
Gets completed assignments for a HIT. You can use this operation to retrieve the results for a HIT

Retrieves the amounts of bonuses you have paid to Workers for a given HIT or assignment.
}
