% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_hits.R
\name{list_hits}
\alias{list_hits}
\title{List HITs}
\usage{
list_hits(next_token, max_results)
}
\arguments{
\item{next_token}{Pagination token}

\item{max_results}{Maximum number of Assignments to return}
}
\value{
If successful, a list of HIT objects
}
\description{
Lists all of a Requester's HITs. The operation returns HITs of any status, except for HITs that have been deleted of with the DeleteHIT operation or that have been auto-deleted
}
