% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_hits_for_qualification_type.R
\name{list_hits_for_qualification_type}
\alias{list_hits_for_qualification_type}
\title{List HITs for Qualification Type}
\usage{
list_hits_for_qualification_type(id, next_token, max_results)
}
\arguments{
\item{id}{The id of the Qualification type to use when querying HITs.}

\item{next_token}{Pagination token}

\item{max_results}{Maximum number of Assignments to return}
}
\value{
If successful, a list of HIT objects
}
\description{
Lists HITs that use the given Qualification Type
}
