% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_qualification_requests.R,
%   R/list_qualification_types.R
\name{list_qualification_requests}
\alias{list_qualification_requests}
\title{List Qualification Requests}
\usage{
list_qualification_requests(
  query,
  must_be_requestable,
  must_be_owned_by_caller,
  next_token,
  max_results
)

list_qualification_requests(
  query,
  must_be_requestable,
  must_be_owned_by_caller,
  next_token,
  max_results
)
}
\arguments{
\item{query}{A search term}

\item{must_be_requestable}{Specifies that only Qualification types that a user can request through the Amazon Mechanical Turk web site, such as by taking a Qualification test, are returned as results of the search. Some Qualification types, such as those assigned automatically by the system, cannot be requested directly by users. If false, all Qualification types, including those managed by the system, are considered for the search.}

\item{must_be_owned_by_caller}{Specifies that only Qualification types that the Requester created are returned. If false, the operation returns all Qualification types.}

\item{next_token}{Pagination token}

\item{max_results}{Maximum number of Qualification Types to return}

\item{id}{The id of the Qualification Type}
}
\value{
If successful, a list of Qualification Requests

If successful, a list of Qualification Types
}
\description{
Lists requests for Qualifications of a particular Qualification type

Searches for Qualification types using the specified search query, and returns a list of Qualification types.
}
