% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_review_policy_results_for_hit.R,
%   R/list_reviewable_hits.R
\name{list_review_policy_results_for_hit}
\alias{list_review_policy_results_for_hit}
\title{List Review Policy Results for HIT}
\usage{
list_review_policy_results_for_hit(id, status, next_token, max_results)

list_review_policy_results_for_hit(id, status, next_token, max_results)
}
\arguments{
\item{id}{The id of the HIT type of the HITs to consider for the query}

\item{status}{The status of the HITs to return: Reviewable | Reviewing}

\item{next_token}{Pagination token}

\item{max_results}{Maximum number of HITs to return}

\item{policy_level}{The Policy Level(s) to retrieve review results for - HIT or Assignment. If omitted, the default behavior is to retrieve all data for both policy levels}

\item{assignment}{If supplied, the results are limited to those pertaining directly to this Assignment ID.}

\item{retrieve_actions}{Retrieves a list of the actions taken executing the Review Policies and their outcomes. T or F.}

\item{retrieve_results}{Retrieves a list of the results computed by the Review Policies. T or F.}
}
\value{
If successful, a ListReviewPolicyResultsForHITResponse element containing the name of the Review Policy applied as well as the AssignmentReviewReport element and the HITReviewReport element.

If successful, a list of HIT objects
}
\description{
Lists the computed results and the actions taken in the course of executing your Review Policies during a CreateHIT operation

Retrieves the HITs with Status equal to Reviewable or Status equal to Reviewing that belong to the Requester calling the operation
}
