% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_worker_blocks.R
\name{list_worker_blocks}
\alias{list_worker_blocks}
\title{List Worker Blocks}
\usage{
list_worker_blocks(next_token, max_results)
}
\arguments{
\item{next_token}{Pagination token}

\item{max_results}{Maximum number of Workers to return}
}
\value{
If successful, a list of Workers that have been blocked along with the reason for the block.
}
\description{
Retrieves a list of Workers who are blocked from working on your HITs.
}
