% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_workers_with_qualification_type.R
\name{list_workers_with_qualification_type}
\alias{list_workers_with_qualification_type}
\title{List Workers with Qualification Type}
\usage{
list_workers_with_qualification_type(id, status, next_token, max_results)
}
\arguments{
\item{id}{The ID of the Qualification Type}

\item{status}{The status of the Qualifications to return. Granted | Revoked}

\item{next_token}{Pagination token}

\item{max_results}{Maximum number of Workers to return}
}
\value{
If successful, a list of Qualifications that have been granted to Workers.
}
\description{
Lists all of the Workers with a given Qualification type.
}
