% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify_workers.R
\name{notify_workers}
\alias{notify_workers}
\title{Notify Workers}
\usage{
notify_workers(subject, message_text, workers)
}
\arguments{
\item{subject}{The subject line of the email message to send. Can include up to 200 characters.}

\item{message_text}{The text of the email message to send. Can include up to 4,096 characters.}

\item{workers}{A vector of WorkerIds to notify. You can notify upto 100 Workers at a time.}
}
\value{
If successful, a response with an empty body and no errors
}
\description{
Sends an email to one or more Workers that you specify with the Worker ID. You can specify up to 100 Worker IDs to send the same message with a single call to the NotifyWorkers operation. The NotifyWorkers operation will send a notification email to a Worker only if you have previously approved or rejected work from the Worker.
}
