% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reject_assignment.R
\name{reject_assignment}
\alias{reject_assignment}
\title{Reject Assignment}
\usage{
reject_assignment(id, feedback)
}
\arguments{
\item{id}{The id of the assignment}

\item{feedback}{A message for the Worker, which the Worker can see in the Status section of the web site.}
}
\value{
If successful, a response with an empty body and no errors
}
\description{
Rejects the results of a completed assignment.
}
