% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reject_qualification_request.R
\name{reject_qualification_request}
\alias{reject_qualification_request}
\title{Reject Qualification Request}
\usage{
reject_qualification_request(id, reason)
}
\arguments{
\item{id}{The id of the assignment}

\item{reason}{A text message explaining why the request was rejected, to be shown to the Worker who made the request.}
}
\value{
If successful, a response with an empty body and no errors
}
\description{
Rejects a user's request for a Qualification.
}
