% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_bonus.R
\name{send_bonus}
\alias{send_bonus}
\title{Send Bonus}
\usage{
send_bonus(worker, assignment, amount, reason, .token)
}
\arguments{
\item{worker}{The id of the Worker being paid the bonus}

\item{assignment}{The id of the assignment for which this bonus is paid.}

\item{amount}{The bonus, specified as a US Dollar amount.}

\item{reason}{A message that explains the reason for the bonus payment. The Worker receiving the bonus can see this message.}

\item{.token}{A unique identifier for this request, which allows you to retry the call on error without granting multiple bonuses}
}
\value{
If successful, a response with an empty body and no errors
}
\description{
Issues a payment of money from your account to a Worker. This payment happens separately from the reward you pay to the Worker when you approve the Worker's assignment.
}
