% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_test_event_notification.R
\name{send_test_event_notification}
\alias{send_test_event_notification}
\title{Send Test Event Notification}
\usage{
send_test_event_notification(notification, test_event_type)
}
\arguments{
\item{notification}{The notification specification to test. This value is identical to the value you would provide to the UpdateNotificationSettings operation when you establish the notification specification for a HIT type.}

\item{test_event_type}{The event to simulate to test the notification specification. This event is included in the test message even if the notification specification does not include the event type. The notification specification does not filter out the test event.}
}
\value{
If successful, a response with an empty body and no errors
}
\description{
Causes Amazon Mechanical Turk to send a notification message as if a HIT event occurred, according to the provided notification specification
}
