% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_expiration_for_hit.R
\name{update_expiration_for_hit}
\alias{update_expiration_for_hit}
\title{Update Expiration for HIT}
\usage{
update_expiration_for_hit(id, expire_at)
}
\arguments{
\item{id}{The HIT to update}

\item{expire_at}{The date and time at which you want the HIT to expire}
}
\value{
If successful, a response with an empty body and no errors
}
\description{
Allows you extend the expiration time of a HIT beyond is current expiration or expire a HIT immediately. You cannot shorten the expiration time so that you're not affecting Workers who have accepted your HIT.
}
