% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hit_review_status.R
\name{update_hit_review_status}
\alias{update_hit_review_status}
\title{Update HIT Review Status}
\usage{
update_hit_review_status(id, revert)
}
\arguments{
\item{id}{The HIT to update}

\item{revert}{Specifies whether to update the HIT Status from Reviewing to Reviewable}
}
\value{
If successful, a response with an empty body and no errors
}
\description{
Toggles the status of a HIT. If the status is Reviewable, this operation updates the status to Reviewing, or reverts a Reviewing HIT back to the Reviewable status.
}
