% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hit_type_of_hit.R
\name{update_hit_type_of_hit}
\alias{update_hit_type_of_hit}
\title{Update HIT Type of HIT}
\usage{
update_hit_type_of_hit(id, type)
}
\arguments{
\item{id}{The HIT to update}

\item{type}{The id of the new HIT type}
}
\value{
If successful, a response with an empty body and no errors
}
\description{
Changes the HITType properties of a HIT
}
