% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_notification_settings.R
\name{update_notification_settings}
\alias{update_notification_settings}
\title{Update Notification Settings}
\usage{
update_notification_settings(id, notification, active)
}
\arguments{
\item{id}{The HIT to update}

\item{notification}{The notification specification for the HIT Type}

\item{active}{Specifies whether notifications are sent for HITs of this HIT type, according to the notification specification. You must specify either the Notification parameter or the Active parameter}
}
\value{
If successful, a response with an empty body and no errors
}
\description{
Creates, updates, disables or re-enables notifications for a HIT type.
}
