% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_qualification_type.R
\name{update_qualification_type}
\alias{update_qualification_type}
\title{Update Qualification Type}
\usage{
update_qualification_type(
  id,
  retry_delay,
  status,
  description,
  test,
  answer_key,
  test_duration,
  auto_granted,
  auto_granted_value
)
}
\arguments{
\item{id}{The id of the Qualification type to update.}

\item{retry_delay}{The amount of time, in seconds, that Workers must wait after requesting a Qualification of the specified Qualification type before they can retry the Qualification request}

\item{status}{The new status of the Qualification type - Active | Inactive}

\item{description}{The new description of the Qualification type.}

\item{test}{The questions for the Qualification test a Worker must answer correctly to obtain a Qualification of this type. If this parameter is specified, TestDurationInSeconds must also be specified.}

\item{answer_key}{The answers to the Qualification test specified in the Test parameter, in the form of an AnswerKey data structure.}

\item{test_duration}{The number of seconds the Worker has to complete the Qualification test, starting from the time the Worker requests the Qualification. This is required if the Test parameter is specified.}

\item{auto_granted}{Specifies whether requests for the Qualification type are granted immediately, without prompting the Worker with a Qualification test.}

\item{auto_granted_value}{The Qualification value to use for automatically granted Qualifications. This parameter is used only if the AutoGranted parameter is true.}
}
\value{
If successful, a Qualification Type object
}
\description{
Modifies the attributes of an existing Qualification type, which is represented by a QualificationType data structure. Only the owner of a Qualification type can modify its attributes.
}
