% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accept_qualification_request.R
\name{accept_qualification_request}
\alias{accept_qualification_request}
\title{Accept Qualification Request}
\usage{
accept_qualification_request(id, value, ...)
}
\value{
Empty body and no errors, if successful.
}
\description{
Accepts a request for qualification
}
