% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hit_with_hit_type.R
\name{create_hit_with_hit_type}
\alias{create_hit_with_hit_type}
\title{Create HIT With HIT Type}
\usage{
create_hit_with_hit_type(
  type,
  question,
  layout_id,
  layout_parameters,
  lifetime,
  max_assignments,
  assignment_review_policy,
  hit_review_policy,
  annotation,
  ...,
  .token
)
}
\arguments{
\item{type}{The HIT type ID.}

\item{question}{The data the person completing the HIT uses to produce the results.}

\item{layout_id}{Allows you to use a pre-existing HIT design with placeholder values and create an additional HIT by providing those values as HITLayoutParameters}

\item{layout_parameters}{If the HITLayoutId is provided, any placeholder values must be filled in with values using the HITLayoutParameter structure.}

\item{lifetime}{An amount of time, in seconds, after which the HIT is no longer available for users to accept. After the lifetime of the HIT elapses, the HIT no longer appears in HIT searches, even if not all of the assignments for the HIT have been accepted.}

\item{max_assignments}{The number of times the HIT can be accepted and completed before the HIT becomes unavailable.}

\item{assignment_review_policy}{The Assignment-level Review Policy applies to the assignments under the HIT. You can specify for Mechanical Turk to take various actions based on the policy.}

\item{hit_review_policy}{The HIT-level Review Policy applies to the HIT. You can specify for Mechanical Turk to take various actions based on the policy.}

\item{annotation}{An arbitrary data field that lets your application attach arbitrary data to the HIT for tracking purposes}

\item{...}{Placeholder for future arguments}

\item{.token}{A unique identifier for this request. Allows you to retry the call on error without creating duplicate HITs}
}
\value{
If successful, a HIT object.
}
\description{
Creates a HIT of a particular HIT Type
}
